requirements.txt
adafruit-circuitpython-fingerprint
Pillow
pyserial
requests